+++ telekit/buildtext/__init__.py
@@ -1,0 +1,7 @@
+from . import styles
+from .styles import Styles
+
+__all__ = [
+    "styles",
+    "Styles"
+]