+++ telekit/example/example_handlers/on_text.py
@@ -1,6 +1,9 @@
 import telebot.types # type: ignore
 import telekit
 
+from telekit.buildtext import Styles
+from telekit.buildtext.styles import *
+
 class OnTextHandler(telekit.Handler):
 
     @classmethod
@@ -36,6 +39,25 @@ class OnTextHandler(telekit.Handler):
         if not age:
             age = "An unknown number of"
 
-        self.chain.sender.set_title(f"Hello {name}!")
-        self.chain.sender.set_message(f"{age} years is a wonderful stage of life!")
+        # Manually:
+        # styles = Styles()
+        # styles.use_html()
+        # styles.set_parse_mode("markdown")
+
+        # Automatically detects `parse_mode` according to the `sender`
+        styles = self.chain.sender.styles
+
+        # Another ways:
+        # print(Bold(age).markdown)
+        # print(Bold(age, parse_mode="html"))
+
+        # Composition:
+        #   Strikethrough(Bold("...") + Italic("..."))
+        #   styles.strike(styles.bold("...") + styles.italic())
+
+        self.chain.sender.set_title(f"Hello {styles.italic(name)}!")
+        self.chain.sender.set_message(
+            f"{styles.bold(age)} years is a wonderful stage of life!\n" 
+            f"{styles.quote(f'(You can customize styles)')}"
+        )
         self.chain.send()