+++ telekit/example/example_server.py
@@ -17,7 +17,7 @@ Hey! Welcome to the Telekit family.
 /help  - custom help implementation + scanning files using `chapters`
 /faq   - extended help page written in Telekit DSL for FAQ pages
 
-# Example message handlers:
+# Example message handlers (+ styles example):
 "Name: {name}. Age: {age}"
 "My name is {name} and I am {age} years old"
 "My name is {name}"